+++ helpers.php
@@ -4,6 +4,7 @@ namespace CloudCreativity\JsonApi;
 
 use CloudCreativity\JsonApi\Exceptions\InvalidJsonException;
 use Neomerx\JsonApi\Exceptions\JsonApiException;
+use Psr\Http\Message\MessageInterface;
 
 if (!function_exists('CloudCreativity\JsonApi\json_decode')) {
     /**
@@ -33,5 +34,26 @@ if (!function_exists('CloudCreativity\JsonApi\json_decode')) {
         }
 
         return $parsed;
+    }
+}
+
+if (!function_exists('CloudCreativity\JsonApi\http_contains_body')) {
+    /**
+     * Does the HTTP message contain body content?
+     *
+     * "The presence of a message-body in a request is signaled by the inclusion of a Content-Length or
+     * Transfer-Encoding header field in the request's message-headers."
+     * https://www.w3.org/Protocols/rfc2616/rfc2616-sec4.html#sec4.3
+     *
+     * @param MessageInterface $message
+     * @return bool
+     */
+    function http_contains_body(MessageInterface $message)
+    {
+        if ($message->hasHeader('Content-Length')) {
+            return true;
+        };
+
+        return $message->hasHeader('Transfer-Encoding');
     }
 }