+++ src/Http/Client/SendsRequestsTrait.php
@@ -1,5 +1,21 @@
 <?php
 
+/**
+ * Copyright 2017 Cloud Creativity Limited
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 namespace CloudCreativity\JsonApi\Http\Client;
 
 use CloudCreativity\JsonApi\Contracts\Object\DocumentInterface;
@@ -12,8 +28,14 @@ use Neomerx\JsonApi\Contracts\Schema\ContainerInterface;
 use Neomerx\JsonApi\Encoder\Parameters\EncodingParameters;
 use Neomerx\JsonApi\Http\Headers\MediaType;
 use Psr\Http\Message\ResponseInterface;
+use function CloudCreativity\JsonApi\http_contains_body;
 use function CloudCreativity\JsonApi\json_decode;
 
+/**
+ * Trait SendsRequestsTrait
+ *
+ * @package CloudCreativity\JsonApi
+ */
 trait SendsRequestsTrait
 {
 
@@ -112,10 +134,14 @@ trait SendsRequestsTrait
 
     /**
      * @param ResponseInterface $response
-     * @return DocumentInterface
+     * @return DocumentInterface|null
      */
     protected function decode(ResponseInterface $response)
     {
+        if (!http_contains_body($response)) {
+            return null;
+        }
+
         $content = (string) $response->getBody();
 
         return new Document(json_decode($content));