+++ CHANGELOG.md
@@ -7,8 +7,9 @@ All notable changes to this project will be documented in this file. This projec
 ### Added
 - Can now get included resource objects from the JSON API document object.
 - Can now get error objects from the JSON API document object.
-- A `json_decode` helper function is now available to decode JSON strings and throw a JSON API exception if
-an error occurs.
+- Added the following helper functions:
+  - `json_decode()`: wraps the native function and throws a JSON API exception if decoding fails.
+  - `http_contains_body()`: determines if a PSR message contains body content.
 
 ### Fixed
 - A decoder is now only matched if the client has sent a HTTP request with message body.